pyproject.toml
[tool.poetry]
name = "idk"
version = "0.1.0"
description = ""
authors = ["Grigory Malanchuk <grigorymalanchuk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
django-filter = "^24.2"
celery = "^5.4.0"
aiohttp = "^3.9.5"
django-redis = "^5.4.0"
flower = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
